requirements.txt
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.4
idna==3.11
lxml==6.0.4
markdown-it-py==4.0.0
mdurl==0.1.2
Pygments==2.19.2
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.32.5
rich==14.3.3
soupsieve==2.8.3
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
